Chunk dimensions in the confi are now shot -> int

+++ Chunky.Shared/ChunkyConfig.cs
@@ -37,11 +37,11 @@ namespace Chunky.Shared
         /// <summary>
         /// The width of each of the chunks. Overrides ChunkCountX if that is set in this config.
         /// </summary>
-        public short ChunkWidth { get; set; }
+        public int ChunkWidth { get; set; }
         /// <summary>
         /// The height of each of the chunks. Overrides ChunkCountY if that is set in this config.
         /// </summary>
-        public short ChunkHeight { get; set; }
+        public int ChunkHeight { get; set; }
         
         // Can't have both these and ChunkWidth/ChunkHeight. Width and Height override CountX and CountY.
         /// <summary>
